[iOS] [12590] Working on `GeneralScheduleFilterViewController`
