--- OpenStack Summit/OpenStack Summit/GeneralScheduleFilterTableViewCell.xib
@@ -1,76 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<document type="com.apple.InterfaceBuilder3.CocoaTouch.XIB" version="3.0" toolsVersion="11542" systemVersion="16B2553a" targetRuntime="iOS.CocoaTouch" propertyAccessControl="none" useAutolayout="YES" useTraitCollections="YES" colorMatched="YES">
-    <device id="retina4_0" orientation="portrait">
-        <adaptation id="fullscreen"/>
-    </device>
-    <dependencies>
-        <deployment identifier="iOS"/>
-        <plugIn identifier="com.apple.InterfaceBuilder.IBCocoaTouchPlugin" version="11524"/>
-        <capability name="documents saved in the Xcode 8 format" minToolsVersion="8.0"/>
-    </dependencies>
-    <objects>
-        <placeholder placeholderIdentifier="IBFilesOwner" id="-1" userLabel="File's Owner"/>
-        <placeholder placeholderIdentifier="IBFirstResponder" id="-2" customClass="UIResponder"/>
-        <tableViewCell contentMode="scaleToFill" selectionStyle="default" indentationWidth="10" reuseIdentifier="GeneralScheduleFilterTableViewCell" rowHeight="52" id="KGk-i7-Jjw" customClass="GeneralScheduleFilterTableViewCell" customModule="OpenStackSummit" customModuleProvider="target">
-            <rect key="frame" x="0.0" y="0.0" width="560" height="52"/>
-            <autoresizingMask key="autoresizingMask" flexibleMaxX="YES" flexibleMaxY="YES"/>
-            <tableViewCellContentView key="contentView" opaque="NO" clipsSubviews="YES" multipleTouchEnabled="YES" contentMode="center" tableViewCell="KGk-i7-Jjw" id="H2p-sc-9uM">
-                <rect key="frame" x="0.0" y="0.0" width="560" height="51"/>
-                <autoresizingMask key="autoresizingMask"/>
-                <subviews>
-                    <view autoresizesSubviews="NO" opaque="NO" clearsContextBeforeDrawing="NO" userInteractionEnabled="NO" contentMode="scaleToFill" translatesAutoresizingMaskIntoConstraints="NO" id="jcz-47-eAb" userLabel="Circle View">
-                        <rect key="frame" x="18" y="13" width="27" height="27"/>
-                        <color key="backgroundColor" red="0.070588235289999995" green="0.1450980392" blue="0.22352941179999999" alpha="1" colorSpace="custom" customColorSpace="sRGB"/>
-                        <accessibility key="accessibilityConfiguration">
-                            <accessibilityTraits key="traits" notEnabled="YES"/>
-                        </accessibility>
-                        <constraints>
-                            <constraint firstAttribute="height" constant="27" id="Yz6-HK-Ixe"/>
-                            <constraint firstAttribute="width" constant="27" id="idv-Yh-jbS"/>
-                        </constraints>
-                        <userDefinedRuntimeAttributes>
-                            <userDefinedRuntimeAttribute type="number" keyPath="layer.cornerRadius">
-                                <real key="value" value="13.5"/>
-                            </userDefinedRuntimeAttribute>
-                            <userDefinedRuntimeAttribute type="boolean" keyPath="layer.masksToBounds" value="YES"/>
-                        </userDefinedRuntimeAttributes>
-                    </view>
-                    <label autoresizesSubviews="NO" opaque="NO" userInteractionEnabled="NO" contentMode="left" verticalHuggingPriority="251" text="Filter Name" textAlignment="natural" lineBreakMode="tailTruncation" numberOfLines="0" baselineAdjustment="alignBaselines" adjustsFontSizeToFit="NO" translatesAutoresizingMaskIntoConstraints="NO" id="HEo-eh-bSR" customClass="CopyableLabel" customModule="OpenStackSummit" customModuleProvider="target">
-                        <rect key="frame" x="18" y="15" width="486" height="22"/>
-                        <fontDescription key="fontDescription" type="system" pointSize="18"/>
-                        <color key="textColor" red="1" green="1" blue="1" alpha="1" colorSpace="custom" customColorSpace="sRGB"/>
-                        <nil key="highlightedColor"/>
-                    </label>
-                    <imageView opaque="NO" userInteractionEnabled="NO" contentMode="scaleToFill" horizontalHuggingPriority="251" verticalHuggingPriority="251" image="check" translatesAutoresizingMaskIntoConstraints="NO" id="Q86-HF-ZbB" userLabel="Check Image">
-                        <rect key="frame" x="522" y="19" width="20" height="14"/>
-                        <constraints>
-                            <constraint firstAttribute="width" constant="20" id="6ma-wb-ieH"/>
-                            <constraint firstAttribute="height" constant="14" id="uMt-Dv-Nbl"/>
-                        </constraints>
-                    </imageView>
-                </subviews>
-                <color key="backgroundColor" red="0.078431372549019607" green="0.15294117647058825" blue="0.23921568627450979" alpha="1" colorSpace="custom" customColorSpace="sRGB"/>
-                <constraints>
-                    <constraint firstItem="Q86-HF-ZbB" firstAttribute="trailing" secondItem="H2p-sc-9uM" secondAttribute="trailing" constant="-18" id="1P8-CI-AH3"/>
-                    <constraint firstItem="Q86-HF-ZbB" firstAttribute="centerY" secondItem="HEo-eh-bSR" secondAttribute="centerY" id="3Sn-PE-8z3"/>
-                    <constraint firstItem="HEo-eh-bSR" firstAttribute="centerY" secondItem="H2p-sc-9uM" secondAttribute="centerY" id="QZk-dG-hx8"/>
-                    <constraint firstItem="HEo-eh-bSR" firstAttribute="leading" secondItem="H2p-sc-9uM" secondAttribute="leading" constant="18" id="ftZ-uf-wg5"/>
-                    <constraint firstItem="Q86-HF-ZbB" firstAttribute="leading" secondItem="HEo-eh-bSR" secondAttribute="trailing" constant="18" id="kii-ic-My9"/>
-                    <constraint firstItem="jcz-47-eAb" firstAttribute="centerY" secondItem="HEo-eh-bSR" secondAttribute="centerY" id="ric-5d-ZkC"/>
-                    <constraint firstItem="jcz-47-eAb" firstAttribute="leading" secondItem="H2p-sc-9uM" secondAttribute="leading" constant="18" id="xJq-eJ-z02"/>
-                </constraints>
-            </tableViewCellContentView>
-            <connections>
-                <outlet property="circleView" destination="jcz-47-eAb" id="NcW-fo-0sZ"/>
-                <outlet property="nameLabel" destination="HEo-eh-bSR" id="Gb5-T6-PcH"/>
-                <outlet property="nameLabelLeadingConstraint" destination="ftZ-uf-wg5" id="CnN-5I-lR3"/>
-                <outlet property="nameLabelVerticalConstraint" destination="QZk-dG-hx8" id="9Pc-Bk-5Q7"/>
-                <outlet property="optionSelectedImage" destination="Q86-HF-ZbB" id="mUt-7v-EC3"/>
-            </connections>
-            <point key="canvasLocation" x="576" y="393"/>
-        </tableViewCell>
-    </objects>
-    <resources>
-        <image name="check" width="34" height="24"/>
-    </resources>
-</document>
