+++ OpenStack Summit/OpenStack Summit/TableViewHeaderViewLight.xib
@@ -1,0 +1,54 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<document type="com.apple.InterfaceBuilder3.CocoaTouch.XIB" version="3.0" toolsVersion="11542" systemVersion="16D32" targetRuntime="iOS.CocoaTouch" propertyAccessControl="none" useAutolayout="YES" useTraitCollections="YES" colorMatched="YES">
+    <device id="retina4_7" orientation="portrait">
+        <adaptation id="fullscreen"/>
+    </device>
+    <dependencies>
+        <plugIn identifier="com.apple.InterfaceBuilder.IBCocoaTouchPlugin" version="11524"/>
+        <capability name="documents saved in the Xcode 8 format" minToolsVersion="8.0"/>
+    </dependencies>
+    <customFonts key="customFonts">
+        <array key="OpenSans-Semibold.ttf">
+            <string>OpenSans-Semibold</string>
+        </array>
+    </customFonts>
+    <objects>
+        <placeholder placeholderIdentifier="IBFilesOwner" id="-1" userLabel="File's Owner"/>
+        <placeholder placeholderIdentifier="IBFirstResponder" id="-2" customClass="UIResponder"/>
+        <view contentMode="scaleToFill" id="jiU-9N-9do" customClass="TableViewHeaderView" customModule="OpenStackSummit" customModuleProvider="target">
+            <rect key="frame" x="0.0" y="0.0" width="320" height="59"/>
+            <autoresizingMask key="autoresizingMask" flexibleMaxX="YES" flexibleMaxY="YES"/>
+            <subviews>
+                <label opaque="NO" userInteractionEnabled="NO" contentMode="left" horizontalHuggingPriority="251" verticalHuggingPriority="251" text="SECTION TITLE" textAlignment="natural" lineBreakMode="tailTruncation" baselineAdjustment="alignBaselines" adjustsFontSizeToFit="NO" translatesAutoresizingMaskIntoConstraints="NO" id="y9z-Z3-QsV">
+                    <rect key="frame" x="16" y="30" width="91" height="18"/>
+                    <fontDescription key="fontDescription" name="OpenSans-Semibold" family="Open Sans" pointSize="13"/>
+                    <color key="textColor" white="0.66666666666666663" alpha="1" colorSpace="calibratedWhite"/>
+                    <nil key="highlightedColor"/>
+                </label>
+                <view contentMode="scaleToFill" translatesAutoresizingMaskIntoConstraints="NO" id="jOI-ex-kSZ" userLabel="Separator View">
+                    <rect key="frame" x="0.0" y="58" width="320" height="1"/>
+                    <color key="backgroundColor" red="0.90196078430000004" green="0.90196078430000004" blue="0.90196078430000004" alpha="1" colorSpace="custom" customColorSpace="sRGB"/>
+                    <constraints>
+                        <constraint firstAttribute="height" constant="1" id="T3z-uh-f7i"/>
+                    </constraints>
+                </view>
+            </subviews>
+            <color key="backgroundColor" white="1" alpha="1" colorSpace="calibratedWhite"/>
+            <constraints>
+                <constraint firstAttribute="trailing" relation="greaterThanOrEqual" secondItem="y9z-Z3-QsV" secondAttribute="trailing" constant="16" id="BrX-vl-GVo"/>
+                <constraint firstItem="y9z-Z3-QsV" firstAttribute="top" secondItem="jiU-9N-9do" secondAttribute="top" constant="30" id="JUs-bi-UaL"/>
+                <constraint firstAttribute="bottom" relation="greaterThanOrEqual" secondItem="y9z-Z3-QsV" secondAttribute="bottom" constant="10" id="PR7-EG-vAS"/>
+                <constraint firstAttribute="trailing" secondItem="jOI-ex-kSZ" secondAttribute="trailing" id="SFx-dq-Tbe"/>
+                <constraint firstItem="jOI-ex-kSZ" firstAttribute="leading" secondItem="jiU-9N-9do" secondAttribute="leading" id="SWK-nN-xvc"/>
+                <constraint firstItem="y9z-Z3-QsV" firstAttribute="leading" secondItem="jiU-9N-9do" secondAttribute="leading" constant="16" id="jSp-oZ-6If"/>
+                <constraint firstAttribute="bottom" secondItem="jOI-ex-kSZ" secondAttribute="bottom" id="jZ5-Jr-JIb"/>
+            </constraints>
+            <nil key="simulatedStatusBarMetrics"/>
+            <freeformSimulatedSizeMetrics key="simulatedDestinationMetrics"/>
+            <connections>
+                <outlet property="titleLabel" destination="y9z-Z3-QsV" id="lhL-yN-A3f"/>
+            </connections>
+            <point key="canvasLocation" x="-1183" y="-431.5"/>
+        </view>
+    </objects>
+</document>
